arena, physics fix, bg tiles

+++ Breakout/Vector2.h
@@ -4,15 +4,15 @@
 
 struct Vector2
 {
-	int x;
-	int y;
+	float x;
+	float y;
 
 	Vector2() { x = 0; y = 0; }
-	Vector2(int X, int Y) { x = X; y = Y; }
+	Vector2(float X, float Y) { x = X; y = Y; }
 
 	static Vector2 Zero() { return Vector2(0, 0); }
 
-	static int DistanceSqr(Vector2 A, Vector2 B) { return pow(A.x - B.x, 2) + pow(A.y - B.y, 2); }
+	static float DistanceSqr(Vector2 A, Vector2 B) { return pow(A.x - B.x, 2) + pow(A.y - B.y, 2); }
 	static Vector2 minDistance(Vector2 Origin, Vector2 A, Vector2 B) { return DistanceSqr(Origin, A) < DistanceSqr(Origin, B) ? A : B; }
 
 	static float toRadians(int degrees)
@@ -20,17 +20,17 @@ struct Vector2
 		return degrees * (M_PI / 180.0f);
 	}
 
-	static int toDegrees(float radians)
+	static float toDegrees(float radians)
 	{
 		return radians * (180.0f / M_PI);
 	}
 
-	static int Length(Vector2 vec)
+	static float Length(Vector2 vec)
 	{
 		return sqrt(vec.x * vec.x + vec.y * vec.y);
 	}
 
-	static int VecToAngle(const Vector2& vectorDistanceToOther)
+	static float VecToAngle(const Vector2& vectorDistanceToOther)
 	{
 		float rads = std::atan2(vectorDistanceToOther.y , vectorDistanceToOther.x);
 		return toDegrees(rads);
@@ -48,7 +48,7 @@ struct Vector2
 		return result;
 	}
 
-	Vector2& Multiply(int X)
+	Vector2& Multiply(float X)
 	{
 		x *= X;
 		y *= X;
